# trie: parallel insert trie when root node is a full node